+++ Assets/FbxExporters/Editor/ConvertToPrefabEditorWindow.cs
@@ -83,6 +83,19 @@ namespace FbxExporters
                 m_prefabExtLabelWidth = m_fbxExtLabelStyle.CalcSize (new GUIContent (".prefab")).x;
             }
 
+            protected bool ExportSetContainsAnimation ()
+            {
+                foreach(var obj in ToExport)
+                {
+                    var go = ModelExporter.GetGameObject(obj);
+                    if(go.GetComponentInChildren<Animation>() || go.GetComponentInChildren<Animator>())
+                    {
+                        return true;
+                    }
+                }
+                return false;
+            }
+
             protected override bool Export ()
             {
                 if (string.IsNullOrEmpty (m_exportFileName)) {
@@ -104,6 +117,29 @@ namespace FbxExporters
                 if (ToExport == null) {
                     Debug.LogError ("FbxExporter: missing object for conversion");
                     return false;
+                }
+
+                if (SettingsObject.UseMayaCompatibleNames && SettingsObject.AllowSceneModification)
+                {
+                    string warning = "Names of objects in the hierarchy may change with the Compatible Naming option turned on";
+                    if (ExportSetContainsAnimation())
+                    {
+                        warning = "Compatible Naming option turned on. Names of objects in hierarchy may change and break animations.";
+                    }
+
+                    // give a warning dialog that indicates that names in the scene may change
+                    int result = UnityEditor.EditorUtility.DisplayDialogComplex(
+                                    string.Format("{0} Warning", ModelExporter.PACKAGE_UI_NAME), warning, "OK", "Turn off and continue", "Cancel"
+                                );
+                    if (result == 1)
+                    {
+                        // turn compatible naming off
+                        SettingsObject.SetUseMayaCompatibleNames(false);
+                    }
+                    else if (result == 2)
+                    {
+                        return false;
+                    }
                 }
 
                 if (ToExport.Length == 1) {